pyproject.toml
[project]
name = "remnashop"
version = "0.0.0"
description = "This is a Telegram bot for selling VPN subscriptions. It works with Remnawave"
authors = [{ name = "snoups" }]
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "aiogram~=3.22.0",
    "aiogram-dialog~=2.4.0",
    "alembic~=1.16.5",
    "asyncpg~=0.30.0",
    "dishka~=1.6.0",
    "fastapi>=0.120.2",
    "fluentogram~=1.2.1",
    "greenlet>=3.2.4",
    "httpx>=0.27.2,<0.28.0",
    "loguru~=0.7.3",
    "msgspec~=0.19.0",
    "pydantic>=2.4.1,<2.12",
    "pydantic-settings~=2.11.0",
    "redis~=6.4.0",
    "remnawave>=2.3.2",
    #"remnawave @ file:///opt/python-sdk/remnawave",
    #"remnawave @ git+https://github.com/remnawave/python-sdk.git@development",
    "taskiq~=0.11.19",
    "taskiq-redis~=1.1.2",
    "uvicorn>=0.38.0",
    "cryptography>=46.0.3",
    "bcrypt>=4.0.0",
    "sqlalchemy[mypy]>=2.0.0",
    "qrcode[pil]>=8.2",
    "python-jose[cryptography]>=3.4.0",
    "aiofiles>=24.1.0",
    "pillow>=11.0.0",
]

[project.urls]
homepage = "https://t.me/@remna_shop"
repository = "https://github.com/snoups/remnashop"

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "ruff>=0.14.2",
    "pytest>=8.4.2",
    "watchfiles>=1.1.1",
    "ftl-extract>=0.9.0",
    "types-cachetools",
]

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = [
    "venv",
    ".venv",
    ".idea",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    "web-app/node_modules",
    "web-app/dist",
]

[tool.ruff.lint]
select = [
    "C",
    "DTZ",
    "E",
    "F",
    "I",
    "ICN",
    "N",
    "PLC",
    "PLE",
    "Q",
    "T",
    "W",
    "YTT",
]
ignore = ["N805"]
[tool.ruff.lint.per-file-ignores]
"src/infrastructure/database/migrations/versions/0023_add_currency_and_gateway_types.py" = ["E501"]
"src/infrastructure/database/migrations/versions/0025_add_partner_settings_column.py" = ["E501"]
"src/infrastructure/database/migrations/versions/0026_add_partner_individual_settings.py" = ["E501"]
"src/infrastructure/database/migrations/versions/0034_add_branding_settings.py" = ["E501"]

[tool.mypy]
packages = ["src"]
plugins = ["sqlalchemy.ext.mypy.plugin", "pydantic.mypy"]
exclude = ["venv", ".venv", ".idea", "tests"]
warn_unused_configs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_reexport = true
strict_equality = true
extra_checks = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = ["fluentogram.*", "remnawave.*", "qrcode.*", "jose.*", "aiofiles.*"]
follow_untyped_imports = true

[[tool.mypy.overrides]]
module = ["app.bot.handlers.*", "src.bot.routers.*", "src.infrastructure.payment_gateways.*"]
strict_optional = false
warn_return_any = false
disable_error_code = ["union-attr", "arg-type", "assignment"]

[[tool.mypy.overrides]]
module = ["src.services.backup", "src.services.partner", "src.infrastructure.taskiq.tasks.*"]
disable_error_code = ["call-overload", "unused-ignore"]
